backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-dotenv==1.1.1
requests==2.31.0
pydantic==2.11.7
google-generativeai==0.8.5
